nastavenie pouzivanie enteru pri mazani (t.j. ak stlacime enter pri pytani, ci chceme mazat, tak enter znamena ano)

+++ LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteItemController.java
@@ -4,6 +4,9 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import sk.upjs.paz1c.business.UserIdentificationManager;
 import sk.upjs.paz1c.entities.Item;
 import sk.upjs.paz1c.entities.Project;
 import sk.upjs.paz1c.fxmodels.ItemFxModel;
@@ -13,31 +16,30 @@ import sk.upjs.paz1c.persistent.ItemDAO;
 import sk.upjs.paz1c.persistent.ProjectDAO;
 
 public class DeleteItemController {
-	
+
 	@FXML
 	private Button yesButton;
 
 	@FXML
 	private Button noButton;
-	
+
 	private ItemDAO itemDao = DAOfactory.INSTANCE.getItemDAO();
-	private ItemFxModel itemModel; 
-	
+	private ItemFxModel itemModel;
+
 	public DeleteItemController(Item item) {
-    	this.itemModel = new ItemFxModel(item);
+		this.itemModel = new ItemFxModel(item);
 	}
-	
+
 	@FXML
 	void initialize() {
-		
-		
-		yesButton.setOnAction(new EventHandler<ActionEvent>() {
+		yesButton.setOnKeyPressed(new EventHandler<KeyEvent>() {
 
 			@Override
-			public void handle(ActionEvent event) {
-				itemDao.deleteItem(itemModel.getItem());
-				//projectDao.saveProject(projectModel.getProject());
-				yesButton.getScene().getWindow().hide();
+			public void handle(KeyEvent event) {
+				if (event.getCode().equals(KeyCode.ENTER)) {
+					itemDao.deleteItem(itemModel.getItem());
+					yesButton.getScene().getWindow().hide();
+				}
 			}
 		});
