+++ LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteNoteController.java
@@ -4,6 +4,9 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import sk.upjs.paz1c.business.UserIdentificationManager;
 import sk.upjs.paz1c.entities.Note;
 import sk.upjs.paz1c.entities.Task;
 import sk.upjs.paz1c.fxmodels.NoteFxModel;
@@ -14,30 +17,31 @@ import sk.upjs.paz1c.persistent.TaskDAO;
 
 public class DeleteNoteController {
 	private NoteDAO noteDao = DAOfactory.INSTANCE.getNoteDAO();
-	private NoteFxModel noteModel; 
-	private Note note; 
+	private NoteFxModel noteModel;
+	private Note note;
 
 	@FXML
-    private Button yesButton;
+	private Button yesButton;
 
-    @FXML
-    private Button noButton;
-    
-    public DeleteNoteController(Note note) {
-		this.note = note; 
-    	this.noteModel = new NoteFxModel(note);
+	@FXML
+	private Button noButton;
+
+	public DeleteNoteController(Note note) {
+		this.note = note;
+		this.noteModel = new NoteFxModel(note);
 	}
-    
+
 	@FXML
 	void initialize() {
-		
-		
-		yesButton.setOnAction(new EventHandler<ActionEvent>() {
+
+		yesButton.setOnKeyPressed(new EventHandler<KeyEvent>() {
 
 			@Override
-			public void handle(ActionEvent event) {
-				noteDao.deleteNote(noteModel.getNote());
-				yesButton.getScene().getWindow().hide();
+			public void handle(KeyEvent event) {
+				if (event.getCode().equals(KeyCode.ENTER)) {
+					noteDao.deleteNote(noteModel.getNote());
+					yesButton.getScene().getWindow().hide();
+				}
 			}
 		});
