+++ LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteProjectController.java
@@ -4,16 +4,19 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import sk.upjs.paz1c.business.UserIdentificationManager;
 import sk.upjs.paz1c.entities.Project;
 import sk.upjs.paz1c.fxmodels.ProjectFxModel;
 import sk.upjs.paz1c.persistent.DAOfactory;
 import sk.upjs.paz1c.persistent.ProjectDAO;
 
 public class DeleteProjectController {
-	
+
 	private ProjectDAO projectDao = DAOfactory.INSTANCE.getProjectDAO();
-	private ProjectFxModel projectModel; 
-	private Project project; 
+	private ProjectFxModel projectModel;
+	private Project project;
 
 	@FXML
 	private Button yesButton;
@@ -22,21 +25,21 @@ public class DeleteProjectController {
 	private Button noButton;
 
 	public DeleteProjectController(Project project) {
-		this.project = project; 
-    	this.projectModel = new ProjectFxModel(project);
+		this.project = project;
+		this.projectModel = new ProjectFxModel(project);
 	}
-	
+
 	@FXML
 	void initialize() {
-		
-		
-		yesButton.setOnAction(new EventHandler<ActionEvent>() {
+
+		yesButton.setOnKeyPressed(new EventHandler<KeyEvent>() {
 
 			@Override
-			public void handle(ActionEvent event) {
-				projectDao.deleteProject(projectModel.getProject());
-				//projectDao.saveProject(projectModel.getProject());
-				yesButton.getScene().getWindow().hide();
+			public void handle(KeyEvent event) {
+				if (event.getCode().equals(KeyCode.ENTER)) {
+					projectDao.deleteProject(projectModel.getProject());
+					yesButton.getScene().getWindow().hide();
+				}
 			}
 		});
