+++ LabBook/src/main/java/sk/upjs/paz1c/gui/DeleteTaskController.java
@@ -4,6 +4,9 @@ import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import sk.upjs.paz1c.business.UserIdentificationManager;
 import sk.upjs.paz1c.entities.Project;
 import sk.upjs.paz1c.entities.Task;
 import sk.upjs.paz1c.fxmodels.ProjectFxModel;
@@ -12,33 +15,34 @@ import sk.upjs.paz1c.persistent.DAOfactory;
 import sk.upjs.paz1c.persistent.TaskDAO;
 
 public class DeleteTaskController {
-	
+
 	private TaskDAO taskDao = DAOfactory.INSTANCE.getTaskDAO();
-	private TaskFxModel taskModel; 
-	private Task task; 
+	private TaskFxModel taskModel;
+	private Task task;
 
 	@FXML
-    private Button yesButton;
+	private Button yesButton;
 
-    @FXML
-    private Button noButton;
-    
-    public DeleteTaskController(Task task) {
-		this.task = task; 
-    	this.taskModel = new TaskFxModel(task);
+	@FXML
+	private Button noButton;
+
+	public DeleteTaskController(Task task) {
+		this.task = task;
+		this.taskModel = new TaskFxModel(task);
 	}
-    
+
 	@FXML
 	void initialize() {
-		
-		
-		yesButton.setOnAction(new EventHandler<ActionEvent>() {
+
+		yesButton.setOnKeyPressed(new EventHandler<KeyEvent>() {
 
 			@Override
-			public void handle(ActionEvent event) {
-				taskDao.deleteTask(taskModel.getTask());
-				taskDao.saveTask(taskModel.getTask());
-				yesButton.getScene().getWindow().hide();
+			public void handle(KeyEvent event) {
+				if (event.getCode().equals(KeyCode.ENTER)) {
+					taskDao.deleteTask(taskModel.getTask());
+					taskDao.saveTask(taskModel.getTask());
+					yesButton.getScene().getWindow().hide();
+				}
 			}
 		});
